Added setting BULK to database configuration
BULK is the number of records that may be bulk inserted into a database
table. SQLite has a weird restriction on the number of "variables" that
can be inserted during bulk insert. PostgreSQL, on the other hand, has
no such limitation. Using 999 as a reasonable number of rows to insert
at a time.