Settings file being added after sanitization
The settings file was inadvertently committed with passwords in it. Used
BFG Repo-Cleaner to clean the history after the file was removed from
the repository. Adding the file back without the passwords. Although we
lost the hisroty of the earlier file, it was not worth the risk of
having the database compromised.