Safe-guard to prevent loading non-text files
Prior to the introduction of the safe-guard, an implicit assuption, that
all files in the gprof directory will be text files was made. As a
result, non-text files in the gprof directory could cause the loader to
crash with a decoding error. The safe-guard ensures only text files in
the gprof directory are considered for loading.