Bug: Backported fixes considered as future fixes
The root cause of the bug was the way in which one release was being
compared to another. Before this change, the only criteria considered to
tell if a release is before or after a given release was the date of the
release. Although the date of a release seems logical, it fails when a
release containing the backports to vulnerability fixes is made after
the date of the release in which the vulnerabilities were found. For
example, consider a release, say 2.5.0, that contains fixes for 10
vulnerabilities. 2.5.0 was released on 01/01/2016, however, the fixes to
the 10 vulnerabilities were backported to the 2.4.x version branch and
released as 2.4.11 on 01/10/2016. If we were to rely only on the date,
the fixes in 2.4.11 would be miscontrued as fixes for future
vulnerabilities. With the fix, the comparison now takes version number
into consideration as well.